Test code

+++ AzentioAutomationFramework_AutoLoan/src/test/java/tests/TestB.java
@@ -3,7 +3,7 @@ package tests;
 public class TestB {
 
 	public static void main(String args[]) {
-		int a = 10;
+		double a = 10;
 		TestB test = new TestB();
 		System.out.println("a " + a);
 		TestA testA = new TestA(15);
